split out distributions
