+++ system/common.hpp
@@ -404,7 +404,7 @@ namespace Grappa {
 /// @addtogroup Utility
 /// @{
 
-namespace impl{
+namespace impl {
 
 static inline void prefetcht0(const void *p) {
     __builtin_prefetch(p, 0, 3);
@@ -421,6 +421,17 @@ static inline void prefetcht2(const void *p) {
 }
 
 /// @}
+
+namespace impl {
+
+// Helper function to make it possible to use static assert to throw
+// an error only if a template is instantiated.
+template<typename T> 
+constexpr bool templateStaticAssertFalse() {
+  return false;
+}
+
+}
 
 }
 #endif
